==> notebooks/2 - prepare vulnerabilities and bugs datasets/1 - Prepare version and time lag at update date.ipynb <==
%matplotlib inline
import seaborn as sns; 
import pandas as pd
import matplotlib.pyplot as plt
from matplotlib import style
from matplotlib import gridspec
import scipy
style.use('ggplot')
import datetime
from io import StringIO
import re
import os
import numpy as np
from datetime import datetime
import apt_pkg
apt_pkg.init_system()

# Load installed packages

packages = pd.read_csv('../../data/for_analysis/installed_packages.csv',
                      usecols=['package','version','archive','last_updated','date'])

packages.drop_duplicates(inplace=True)

packages.shape

packages[['package','version','last_updated']].drop_duplicates().shape

for col in packages.columns:
    packages[col] = packages[col].apply(str)

packages.head(2)

# Load debian packages

debian_packages = pd.read_csv('../../data/prepared_data/debian_packages.csv',
                             usecols=['package','version','archive','date'])
debian_packages.drop_duplicates(inplace=True)
debian_packages.shape

for col in debian_packages.columns:
    debian_packages[col] = debian_packages[col].apply(str)

debian_packages.rename(columns={'version':'version_c','date':'date_c'}, inplace=True)

debian_packages.tail(2)

packages = (packages
            .merge(debian_packages,
                   on = ['package','archive'],
                   how = 'left'
                  )
           )
packages.shape

# Remove package version that were created after the containers inspection
packages = packages.query('last_updated>=date_c').copy()
packages.shape

packages.head(2)

# Time lag ~ latest available release

time_lag = packages.query('date_c >= date')
time_lag.shape

time_lag = (time_lag
            .sort_values('date_c', ascending=False)
            .groupby(['package','version','archive','last_updated'])
            .first()
            .reset_index()
           )
time_lag.shape

time_lag['date'] = pd.to_datetime(time_lag['date'])
time_lag['date_c'] = pd.to_datetime(time_lag['date_c'])

time_lag['tlag']=time_lag.apply(lambda d: (d['date_c'] - d['date']).days, axis=1)

time_lag.rename(columns={'version_c':'ideal'}, inplace=True)

time_lag = time_lag[['package','version','archive','last_updated','ideal','tlag']]

time_lag['compare'] = time_lag.apply(lambda d: compare(d['version'], d['ideal']), axis=1)

# Version lag ~ highest available version

def compare(version,compare_to):
    return apt_pkg.version_compare(str(version), str(compare_to))
    # 1 if version > compare_to
        
packages['compare'] = packages.apply(lambda d: compare(d['version'], d['version_c']), axis=1)
packages.shape

packages.compare.unique()

version_lag = packages.query('compare < 0')
version_lag.shape

version_lag = (version_lag[['package','version','archive','last_updated','compare']]
               .groupby(['package','version','archive','last_updated'])
               .count()
               .reset_index()
               .rename(columns = {'compare':'vlag'})
              )
version_lag.shape

# Merge between vlag and tlag

time_lag = (time_lag
            .merge(version_lag,
                  on =['package','version','archive','last_updated'],
                   how = 'left'
                  )
           )
time_lag.fillna(0, inplace=True)
time_lag.shape

time_lag = time_lag.query('vlag==tlag or vlag != 0')

time_lag.to_csv('../../data/for_analysis/tv_lag_ideal.csv', index=False)
